renamed the program to web_crawler.py and the program gets the title of the crawled websites
